% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{count_CDS_reads}
\alias{count_CDS_reads}
\title{count_CDS_reads function to count reads according to CDS info}
\usage{
count_CDS_reads(GeneName, isoform, ribo)
}
\arguments{
\item{GeneName}{Name of gene used}

\item{isoform}{Which isoforms used}

\item{ribo}{riboseq file}
}
\value{
a dataframe of the Ribo-seq counts in chromosome positions
}
\description{
count Ribo-seq reads according to the CDS range
}
