% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{count_uORF_CDS_reads}
\alias{count_uORF_CDS_reads}
\title{count_uORF_CDS_reads function to count reads according to CDS info}
\usage{
count_uORF_CDS_reads(gene, uORF, uORF.isoform = NULL, ribo)
}
\arguments{
\item{gene}{gene ID}

\item{uORF}{uORF ID}

\item{uORF.isoform}{numberic, uORF isoform, default use 1}

\item{ribo}{riboseq file}
}
\value{
a dataframe of the Ribo-seq counts in chromosome positions
}
\description{
count_uORF_CDS_reads count ribo-seq reads in the CDS of a given uORF
}
