% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{firstInFramePSitePerExonNegative}
\alias{firstInFramePSitePerExonNegative}
\title{firstInFramePSitePerExonNegative}
\usage{
firstInFramePSitePerExonNegative(x)
}
\arguments{
\item{x}{Transcript ID}
}
\value{
if the reads are on the negative strand first in frame P-site per exon
}
\description{
if the reads are on the negative strand first in frame P-site per exon
}
