% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{firstInFramePSitePerExonPositive}
\alias{firstInFramePSitePerExonPositive}
\title{firstInFramePSitePerExonPositive}
\usage{
firstInFramePSitePerExonPositive(x)
}
\arguments{
\item{x}{Transcript ID}
}
\value{
if the reads are on the positive strand first in frame P-site per exon
}
\description{
if the reads are on the positive strand first in frame P-site per exon
}
