% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{gene.structure}
\alias{gene.structure}
\title{Output gene structure information.}
\usage{
gene.structure(annotation, format = "gtf", dataSource = "", organism = "")
}
\arguments{
\item{annotation}{Path to the annotation file in either gtf or gff format.}

\item{format}{Is the annotation file in gtf or gff format?}

\item{dataSource}{Which organization annotate the genome? For example, TAIR or Araport for Arabidopsis.}

\item{organism}{Which organism is it? For example, Arabidopsis.}
}
\value{
multiple GRangesLists with transcript, CDS and UTR information.
}
\description{
Take a gtf or gff file to output several GRangesLists for the range information for transcripts, CDS and untranslation regions (UTRs) of the gene of interest to the global environment. These GRangesLists will be used for other functions  of RiboPlotter to define the features of genes.
}
\examples{
\dontrun{
gene.structure(annotation="PathToArabidopsis.gff",format="gff",dataSource="Araport",organism="Arabidopsis thaliana")
}

}
