% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{p_site_plot_genome}
\alias{p_site_plot_genome}
\title{p_site_plot_genome plot RNA-seq and ribo-seq together for one datasets without CDS information}
\usage{
p_site_plot_genome(GeneName, ribo, Extend = Extend, YLIM)
}
\arguments{
\item{GeneName}{Gene ID}

\item{ribo}{riboseq dataset}

\item{Extend}{Integer, plot more at both ends}

\item{YLIM}{Integer, max value of Y-axis}
}
\value{
a plot for the Ribo-seq reads with periodicity
}
\description{
PLOTg plot RNA-seq and ribo-seq together for one datasets. It also contains a plot with transcript models.
}
