% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{p_site_plot_p}
\alias{p_site_plot_p}
\title{p_site_plot_p function to plot periodicity according to transcript info}
\usage{
p_site_plot_p(
  GeneName,
  isoform,
  ribo,
  CDSonly = FALSE,
  Extend = Extend,
  YLIM,
  axesQ,
  Grey = FALSE
)
}
\arguments{
\item{GeneName}{Name of gene used}

\item{isoform}{Which isoforms used}

\item{ribo}{riboseq file}

\item{CDSonly}{TRUE or FALSE}

\item{Extend}{Integer, plot more at both ends}

\item{YLIM}{Integer, max value of Y-axis}

\item{axesQ}{PLOTc or PLOT, determine if the axes will be plotted}

\item{Grey, }{Logical, darkgrey or not for Ribo/Degradome/CAGE reads}
}
\value{
a plot for the Ribo-seq reads with periodicity
}
\description{
Plot Ribo-seq peridicity in three colors according to the frame information.
}
