% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{p_site_plot_p2}
\alias{p_site_plot_p2}
\title{p_site_plot_p2 is for plotting uORFs}
\usage{
p_site_plot_p2(
  gene,
  uORF,
  uORF.isoform = NULL,
  ribo,
  CDSonly = TRUE,
  Extend = Extend,
  YLIM
)
}
\arguments{
\item{gene}{gene ID}

\item{uORF}{uORF ID}

\item{uORF.isoform}{numberic, uORF isoform, default use 1}

\item{ribo}{riboseq file}

\item{CDSonly}{TRUE or FALSE}

\item{Extend}{Integer, plot more at both ends}

\item{YLIM}{Integer, max value of Y-axis}
}
\value{
a plot for the Ribo-seq reads with periodicity
}
\description{
p_site_plot_p2 plots uORFs one at a time and in three colors.
}
