% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{p_site_Y_max}
\alias{p_site_Y_max}
\title{Identify highest value of riboseq reads in a give gene.}
\usage{
p_site_Y_max(GeneName, isoform, ribo, CDSonly = FALSE, Extend = Extend)
}
\arguments{
\item{GeneName}{Name of gene used}

\item{isoform}{Which isoforms used}

\item{ribo}{A data.frame of riboseq reads}

\item{CDSonly}{True or False, do we only want to plot the CDS range}

\item{Extend}{Do we want to plot a wider range. Default is 0.}
}
\value{
The highest value of riboseq reads
}
\description{
Identify highest value of riboseq reads in a give gene. This is for defining the max Y-axis values.
}
