% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{PLOTc}
\alias{PLOTc}
\title{PLOTc plot RNA-seq and ribo-seq together for one datasets.}
\usage{
PLOTc(YFG, data1 = RNAseqBam1, CDSonly = FALSE, Extend = 0, isoform,
  uORF = NULL, NAME = "")
}
\arguments{
\item{YFG}{Gene ID}

\item{data1}{Dataset 1 to plot. Default is RNAseqBam1 that was loaded by rna_bam.ribo.}

\item{CDSonly}{TRUE or FALSE. Only plot CDS region or all riboseq reads in defined area. Default plot all riboseq reads in the defined area.}

\item{Extend}{Integer. The number of extra nt ploted at the ends of the plots.}

\item{isoform}{Integer. Which isoform to plot periodicity.}

\item{uORF}{Gene ID for uORF}

\item{NAME}{Name of the gene}
}
\value{
Both RNAseq and Riboseq plot together for one set of data
}
\description{
PLOTc plot RNA-seq and ribo-seq together for one datasets. It also contains a plot with transcript models.
}
