% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{PLOTch2}
\alias{PLOTch2}
\title{PLOTch2}
\usage{
PLOTch2(
  YFG,
  RNAbam1 = RNAseqBam1,
  RNAbam2 = RNAseqBam2,
  ribo1 = Ribo1,
  ribo2 = Ribo2,
  SAMPLE1 = S_NAME1,
  SAMPLE2 = S_NAME2,
  CDSonly = FALSE,
  Extend = 50,
  isoform,
  uORF = NULL,
  NAME = "",
  uORFisoform
)
}
\arguments{
\item{YFG}{Gene ID}

\item{RNAbam1}{Dataset 1 to plot. Default is RNAseqBam1 that was loaded by rna_bam.ribo.}

\item{RNAbam2}{Dataset 2 to plot. Default is RNAseqBam2 that was loaded by rna_bam.ribo.}

\item{ribo1}{riboseq dataset 1}

\item{ribo2}{riboseq dataset 2}

\item{SAMPLE1}{name of sample 1}

\item{SAMPLE2}{name of sample 2}

\item{CDSonly}{TRUE or FALSE. Only plot CDS region or all riboseq reads in defined area. Default plot all riboseq reads in the defined area.}

\item{Extend}{Integer. The number of extra nt ploted at the ends of the plots.}

\item{isoform}{Integer. Which isoform to plot periodicity.}

\item{uORF}{Gene ID for uORF}

\item{NAME}{Name of the gene}

\item{uORFisoform}{Isoform number of the uORF}
}
\value{
2 plots for RNAseq and Riboseq in 2 different genotypes/conditions.
}
\description{
plot 2 sets of RNA-seq and ribo-seq for comparison. The max height for each plot is according to each sample (PLOTc2 is according to both). It also contains a plot with transcript models.
}
