% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{PLOTch3}
\alias{PLOTch3}
\title{PLOTch3}
\usage{
PLOTch3(
  YFG,
  RNAbam1 = RNAseqBam1,
  ribo1 = Ribo1,
  Deg = DEGdata,
  CAGE = CAGEdata,
  SAMPLE1 = S_NAME1,
  SAMPLE2 = S_NAME2,
  SAMPLE3 = S_NAME3,
  CDSonly = FALSE,
  Extend = 50,
  isoform,
  uORF = NULL,
  NAME = "",
  uORFisoform,
  GREY = c(FALSE, TRUE, TRUE)
)
}
\arguments{
\item{YFG}{Gene ID}

\item{RNAbam1}{Dataset 1 to plot. Default is RNAseqBam1 that was loaded by rna_bam.ribo.}

\item{ribo1}{riboseq dataset 1}

\item{SAMPLE1}{name of sample 3}

\item{CDSonly}{TRUE or FALSE. Only plot CDS region or all riboseq reads in defined area. Default plot all riboseq reads in the defined area.}

\item{Extend}{Integer. The number of extra nt ploted at the ends of the plots.}

\item{isoform}{Integer. Which isoform to plot periodicity.}

\item{uORF}{Gene ID for uORF}

\item{NAME}{Name of the gene}

\item{uORFisoform}{Isoform number of the uORF}

\item{GREY}{grey color or three colors for Ribo/Degradome/CAGE}
}
\value{
3 plots for (1) RNAseq and Riboseq (in 2 different genotypes/conditions. (2) Degradome (3) CAGE
}
\description{
plot 2 sets of RNA-seq and ribo-seq for comparison. It also contains a plot with transcript models.
}
