% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{plotGeneModel}
\alias{plotGeneModel}
\title{plot plotGeneModel}
\usage{
plotGeneModel(gene, uORF, Extend = Extend)
}
\arguments{
\item{gene}{gene ID}

\item{uORF}{uORF ID}

\item{Extend}{number of nucleotides to extend on both side of the gene model}
}
\value{
plot the gene model
}
\description{
plotGeneModel combines both plotRanges and p_site_plot_all functions
}
