% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{plotRanges}
\alias{plotRanges}
\title{plotRanges is a function to plot GRanges}
\usage{
plotRanges(isoform, uORF = NULL, shortest3UTR, ybottom = 0,
  main = deparse(substitute(x)), colCDS = "black", col3 = "white",
  col5 = "lightgrey")
}
\arguments{
\item{isoform}{Integer, the number of isoform}

\item{uORF}{The name of the uORF}

\item{shortest3UTR}{The plotGeneModel will provide this information automatically. Providing the shortest 3UTR range allows this function to decide how to plot the shape of the 3'UTR.}

\item{ybottom}{The ylim for this function.}

\item{main}{Title for the plot. The gene name is the default.}

\item{colCDS}{Color for CDS}

\item{col3}{Color for 3'UTR}

\item{col5}{Color for 5'UTR}
}
\value{
plot each isoform
}
\description{
Plot each isoform based on gtf/gff information
}
