% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{PLOTt}
\alias{PLOTt}
\title{PLOTt, plot RNA-seq and ribo-seq separately.}
\usage{
PLOTt(YFG, RNAbam1 = RNAseqBam1, ribo1 = riboR1, ylab1 = Ribolab1,
  SAMPLE1 = S_NAME1, CDSonly = FALSE, Extend = 50, isoform,
  uORF = NULL, NAME = "")
}
\arguments{
\item{YFG}{Gene ID}

\item{RNAbam1}{Dataset 1 to plot. Default is RNAseqBam1 that was loaded by rna_bam.ribo.}

\item{ribo1}{riboseq dataset}

\item{ylab1}{name of Y-axis}

\item{SAMPLE1}{name of sample 1}

\item{CDSonly}{TRUE or FALSE. Only plot CDS region or all riboseq reads in defined area. Default plot all riboseq reads in the defined area.}

\item{Extend}{Integer. The number of extra nt ploted at the ends of the plots.}

\item{isoform}{Integer. Which isoform to plot periodicity.}

\item{uORF}{Gene ID for uORF}

\item{NAME}{Name of the gene}
}
\value{
One plot with RNA-seq and ribo-seq separately.
}
\description{
Plot both RNA-seq and ribo-seq in two plots with a gene model.
}
