% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{rna_bam.ribo}
\alias{rna_bam.ribo}
\title{Function to get bam file path, ribo-seq p-site information and axis labels information.}
\usage{
rna_bam.ribo(Ribo1, Ribo2 = NULL, RNAseqBam1, RNAseqBam2 = NULL,
  RNAlab1 = "RNA_sample1", RNAlab2 = NULL,
  RNAseqBamPaired = "paired", Ribolab1 = "Ribo_sample1",
  Ribolab2 = NULL, S_NAME1 = "sample1", S_NAME2 = NULL,
  RNAbackground = "#FEFEAE")
}
\arguments{
\item{Ribo1}{The first ribo-seq tsv file path.}

\item{Ribo2}{The second ribo-seq tsv file path.}

\item{RNAseqBam1}{The first RNA-seq bam file path.}

\item{RNAseqBam2}{The second RNA-seq bam file path.}

\item{RNAlab1}{The y-axis label for the first RNA-seq datasets.}

\item{RNAlab2}{The y-axis label for the second RNA-seq datasets.}

\item{RNAseqBamPaired}{Whether the RNA bam is paired-end. Enter "single" for single-end bam file. "paired" for paired-end bam file (default).}

\item{Ribolab1}{The y-axis label for the first ribo-seq datasets.}

\item{Ribolab2}{The y-axis label for the second ribo-seq datasets.}

\item{S_NAME1}{Sample 1 name}

\item{S_NAME2}{Sample 2 name}

\item{RNAbackground}{The background color for RNA-seq results}
}
\value{
Assign pathes or tsv files to global environment required for downstream analysis
}
\description{
This function obtains the RNA-seq bam file path, ribo-seq p-site file information, and the labels for y axes of the RNA-seq and ribo-seq plots
}
\examples{
\dontrun{
uorf.structure(uorf_annotation="/Volumes/BACKUP/project2/TAIR10.29.gtf",dataSource="Araport",organism="Arabidopsis thaliana")
}
}
