% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{rna_bam.ribo3}
\alias{rna_bam.ribo3}
\title{rna_bam.ribo3}
\usage{
rna_bam.ribo3(
  Ribo1,
  DEGdata,
  CAGEdata,
  RNAseqBam1,
  RNAlab1 = "RNA",
  RNAseqBamPaired = "paired",
  Ribolab1 = "Ribo",
  S_NAME1 = "CTRL_Ribo",
  S_NAME2 = "CTRL_Deg",
  S_NAME3 = "CTRL_CAGE",
  RNAbackground = "#FEFEAE"
)
}
\arguments{
\item{Ribo1}{The first ribo-seq tsv file path.}

\item{DEGdata}{The second ribo-seq tsv file path.}

\item{CAGEdata}{The second ribo-seq tsv file path.}

\item{RNAseqBam1}{The first RNA-seq bam file path.}

\item{RNAlab1}{The y-axis label for the first RNA-seq datasets.}

\item{RNAseqBamPaired}{Whether the RNA bam is paired-end. Enter "single" for single-end bam file. "paired" for paired-end bam file (default).}

\item{Ribolab1}{The y-axis label for the first ribo-seq datasets.}

\item{S_NAME1}{Sample 1 name}

\item{S_NAME2}{Sample 2 name}

\item{S_NAME3}{Sample 3 name}

\item{RNAbackground}{The background color for RNA-seq results}
}
\value{
Assign pathes or tsv files to global environment required for downstream analysis
}
\description{
This function obtains the RNA-seq bam file, ribo-seq p-site file, degradome, CAGE file paths and the labels for the y-axis of plots
}
