% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{uorf.structure}
\alias{uorf.structure}
\title{Output gene structure information for uORFs.}
\usage{
uorf.structure(uorf_annotation, format = "gtf", dataSource = "", organism = "")
}
\arguments{
\item{uorf_annotation}{Path to the uORF annotation file in either gtf or gff format.}

\item{format}{Is the uORF annotation file in gtf or gff format?}

\item{dataSource}{Which organization annotate the genome? For example, TAIR or Araport for Arabidopsis.}

\item{organism}{Which organism is it? For example, Arabidopsis.}
}
\value{
Multiple GRangesLists for uORFs with transcript, CDS and UTR information.
}
\description{
uORFs are not usually in the regular annotation files. Therefore you would need to create a gtf or gff of the uORFs of interest. This function takes a gtf or gff file to output several GRangesLists for the range information for transcripts, CDS and untranslation regions (UTRs) of the uORFs to the global environment. These GRangesLists will be used for other functions of RiboPlotter to define the features of the uORF
}
\examples{
\dontrun{
uorf.structure(annotation="PathToArabidopsis.gff",format="gff",dataSource="Araport",organism="Arabidopsis thaliana")
}

}
