% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{PLOTc2}
\alias{PLOTc2}
\title{PLOTc2}
\usage{
PLOTc2(YFG, data1 = RNAseqBam1, data2 = RNAseqBam2, CDSonly = FALSE,
  Extend = 0, isoform, uORF = NULL, Normalized = "", NAME = "")
}
\arguments{
\item{YFG}{Gene ID}

\item{data1}{Data to plot. Default is RNAseqBam1 that was loaded by rna_bam.ribo.}

\item{data2}{Data to plot. Default is RNAseqBam2 that was loaded by rna_bam.ribo.}

\item{CDSonly}{TRUE or FALSE. Only plot CDS region or all riboseq reads in defined area. Default plot all riboseq reads in the defined area.}

\item{Extend}{Integer. The number of extra nt ploted at the ends of the plots.}

\item{isoform}{Integer. Which isoform to plot periodicity.}

\item{uORF}{Gene ID for uORF}

\item{NAME}{Name of the gene}
}
\value{
2 plots for RNAseq and Riboseq in 2 different conditions.
}
\description{
plot 2 sets of RNA-seq and ribo-seq for comparison. It also contains a plot with transcript models.
}
