% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{rna_bam.ribo}
\alias{rna_bam.ribo}
\title{Function to get bam file path, ribo-seq p-site information and axis labels information.}
\usage{
rna_bam.ribo(rna1, ribo1, rna2 = NULL, ribo2 = NULL,
  RNAlab1 = "RNA_sample1", RNAlab2 = "RNA_sample2",
  Ribolab1 = "Ribo_sample1", Ribolab2 = "Ribo_sample2")
}
\arguments{
\item{rna1}{The first RNA-seq bam file path.}

\item{ribo1}{The first ribo-seq tsv file path.}

\item{rna2}{The second RNA-seq bam file path.}

\item{ribo2}{The second ribo-seq tsv file path.}

\item{RNAlab1}{The y-axis label for the first RNA-seq datasets.}

\item{RNAlab2}{The y-axis label for the second RNA-seq datasets.}

\item{Ribolab1}{The y-axis label for the first ribo-seq datasets.}

\item{Ribolab2}{The y-axis label for the second ribo-seq datasets.}
}
\value{
Assign pathes or tsv files to global environment required for downstream analysis
}
\description{
This function obtains the RNA-seq bam file path, ribo-seq p-site file information, and the labels for y axes of the RNA-seq and ribo-seq plots
}
\examples{
\dontrun{
uorf.structure(uorf_annotation="/Volumes/BACKUP/project2/TAIR10.29.gtf",dataSource="Araport",organism="Arabidopsis thaliana")
}
}
